Added t-albumPAge, p-album-page, routing
