+++ .eslintrc.json
@@ -32,6 +32,7 @@
     ],
     "import/prefer-default-export": [
       "off"
-    ]
+    ],
+    "import/no-extraneous-dependencies": ["error", {"devDependencies": true}]
   }
 }
