+++ src/components/_settings/_variables.scss
@@ -1,7 +1,7 @@
 //Colors
 $header-color: #001524;
 $subheader-color: #A0ABB3;
-$background-color: #fff;
+$bg-color: #ffffff;
 $boxShadow-color: 0 8px 20px rgba(240, 242, 245, 0.5);;
 
 //Font size
@@ -14,3 +14,4 @@ $article-header-lineHeight: 21px;
 //Fonts
 $bold: Gilroy, bold, sans-serif;
 $semiBold: Gilroy, semibold, sans-serif;
+
