+++ src/components/atoms/button/button.css
@@ -1,7 +1,7 @@
 .a-button {
-  width: 50%;
+  width: 100%;
   height: 100%;
-  background-color: #fff;
+  background-color: #ffffff;
   font-family: Gilroy, bold, sans-serif;
   font-style: normal;
   font-weight: bold;
