+++ src/components/molecules/Acticle/article.css
@@ -7,7 +7,7 @@
   border-radius: 8px;
   margin-bottom: 0.8em;
   padding: 2.4em;
-  background-color: #fff;
+  background-color: #ffffff;
   box-shadow: 0 8px 20px rgba(240, 242, 245, 0.5);
 }
 .m-article__arrowContainer {
