+++ src/components/molecules/Header/header.scss
@@ -1,5 +1,3 @@
-@use "../../_settings/variables";
-
 .m-header {
   width: 100%;
   height: 8.8em;
