+++ src/components/molecules/Menu/Menu.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import { NavLink } from "react-router-dom";
 import PropTypes from "prop-types";
 import Button from "../../atoms/button/Button";
 import "./menu.css";
@@ -6,21 +7,23 @@ import "../../atoms/button/button.css";
 
 const Menu = ({ ...props }) => {
   const [checked, setChecked] = useState(true);
-  const t = () => setChecked(false);
-  console.log(checked, t);
   return (
     <nav className={props.styles[0]}>
-      <Button
-        styles={!checked ? props.styles[1] : `${props.styles[1]}--active`}
-        onclick={props.firstBtnFunc}
-        title={props.title[0]}
-      />
+      <NavLink to="/news" className="navLink">
+        <Button
+          styles={checked ? props.styles[1] : `${props.styles[1]}--active`}
+          onclick={() => setChecked(true)}
+          title={props.title[0]}
+        />
+      </NavLink>
       <div className="verticalLine" />
-      <Button
-        styles={checked ? props.styles[1] : `${props.styles[1]}--active`}
-        onclick={props.secondBtnFunc}
-        title={props.title[1]}
-      />
+      <NavLink to="/album" className="navLink">
+        <Button
+          styles={!checked ? props.styles[1] : `${props.styles[1]}--active`}
+          onclick={() => setChecked(false)}
+          title={props.title[1]}
+        />
+      </NavLink>
     </nav>
   );
 };
@@ -28,8 +31,6 @@ const Menu = ({ ...props }) => {
 Menu.propTypes = {
   title: PropTypes.arrayOf(PropTypes.string).isRequired,
   styles: PropTypes.arrayOf(PropTypes.string).isRequired,
-  firstBtnFunc: PropTypes.func.isRequired,
-  secondBtnFunc: PropTypes.func.isRequired,
 };
 
 export default Menu;
