+++ src/components/molecules/Menu/menu.css
@@ -11,8 +11,14 @@
 }
 
 .verticalLine {
-  height: 50%;
+  height: 43%;
   border-left: 2px solid #F0F2F5;
+}
+
+.navLink {
+  display: block;
+  width: 50%;
+  height: 100%;
 }
 
 /*# sourceMappingURL=menu.css.map */
