+++ src/components/molecules/Menu/menu.scss
@@ -13,7 +13,12 @@
 }
 
 .verticalLine{
-  height: 50%;
+  height: 43%;
   border-left: 2px solid #F0F2F5;
 }
 
+.navLink{
+  display: block;
+  width: 50%;
+  height: 100%;
+}
