+++ src/components/pages/AlbumPage/AlbumPage.css
@@ -1,0 +1,15 @@
+.header {
+  width: 100%;
+  height: auto;
+}
+
+.section {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 2.4em 0 6.4em 0;
+}
+
+/*# sourceMappingURL=AlbumPage.css.map */
