+++ src/components/templates/AlbumTemplate/albumTemplate.scss
@@ -1,0 +1,14 @@
+.header {
+  width: 100%;
+  height: auto;
+}
+
+.section {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 2.4em 0 6.4em 0;
+  background-color: #fff;
+}
