+++ src/components/templates/NewsListTemplate/newsListTemplate.css
@@ -6,7 +6,6 @@
 .section {
   width: 100%;
   height: 100%;
-  background-color: #F4F6F9;
   display: flex;
   justify-content: center;
   align-items: center;
