+++ src/components/templates/NewsListTemplate/newsListTemplate.scss
@@ -1,4 +1,4 @@
-@use "../_settings/variables";
+@use "../../_settings/variables";
 
 .header {
     width: 100%;
@@ -8,7 +8,6 @@
 .section {
     width: 100%;
     height: 100%;
-    background-color: #F4F6F9;
     display: flex;
     justify-content: center;
     align-items: center;
